requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
python-dotenv
jinja2
motor
pytest
pytest-asyncio
respx
gunicorn
